:art: restructure.
