+++ bindings/pyfiction/include/pyfiction/inout/write_svg_layout.hpp
@@ -1,20 +1,17 @@
 //
-// Created by Jan Drewniok on 05.11.24.
+// Created by marcel on 08.07.22.
 //
 
-#ifndef PYFICTION_WRITE_SIDB_LAYOUT_SVG_HPP
-#define PYFICTION_WRITE_SIDB_LAYOUT_SVG_HPP
+#ifndef PYFICTION_WRITE_SVG_LAYOUT_HPP
+#define PYFICTION_WRITE_SVG_LAYOUT_HPP
 
 #include "pyfiction/documentation.hpp"
 #include "pyfiction/types.hpp"
 
-#include <fiction/io/write_sidb_layout_svg.hpp>
+#include <fiction/io/write_svg_layout.hpp>
 
 #include <pybind11/pybind11.h>
 
-#include <ostream>
-#include <sstream>
-#include <string>
 #include <string_view>
 
 namespace pyfiction
@@ -23,7 +20,7 @@ namespace pyfiction
 namespace detail
 {
 template <typename Lyt>
-inline void write_sidb_layout_svg_impl(pybind11::module& m)
+inline void write_svg_layout_impl(pybind11::module& m)
 {
     using namespace pybind11::literals;
 
@@ -32,13 +29,13 @@ inline void write_sidb_layout_svg_impl(pybind11::module& m)
                                           const fiction::write_sidb_layout_svg_params&) =
         &fiction::write_sidb_layout_svg<Lyt>;
 
-    // Register the original function to write to a file
+    // SiDB plot
     m.def("write_sidb_layout_svg", write_sidb_layout_svg_pointer, "layout"_a, "filename"_a,
           "ps"_a = fiction::write_sidb_layout_svg_params{}, DOC(fiction_write_sidb_layout_svg));
 
     // Register the function to return an SVG as a string
     m.def(
-        "write_sidb_layout_svg_and_return_as_string",
+        "write_sidb_layout_svg_to_string",
         [](const Lyt& layout, const fiction::write_sidb_layout_svg_params& params) -> std::string
         {
             std::ostringstream oss;                               // Create an in-memory output stream
@@ -46,11 +43,21 @@ inline void write_sidb_layout_svg_impl(pybind11::module& m)
             return oss.str();                                     // Return the string content
         },
         "layout"_a, "ps"_a = fiction::write_sidb_layout_svg_params{}, DOC(fiction_write_sidb_layout_svg));
+
+
+    // QCA plot
+    void (*write_qca_layout_svg_pointer)(const py_qca_layout&, const std::string_view&,
+                                              const fiction::write_qca_layout_svg_params&) =
+        &fiction::write_qca_layout_svg<py_qca_layout>;
+
+    m.def("write_qca_layout_svg", write_qca_layout_svg_pointer, "layout"_a, "filename"_a,
+          "params"_a = fiction::write_qca_layout_svg_params{}, DOC(fiction_write_qca_layout_svg));
 }
 
 }  // namespace detail
 
-void write_sidb_layout_svg(pybind11::module& m)
+
+void write_svg_layout(pybind11::module& m)
 {
     namespace py = pybind11;
     using namespace pybind11::literals;
@@ -71,15 +78,22 @@ void write_sidb_layout_svg(pybind11::module& m)
         .def_readwrite("color_background", &fiction::write_sidb_layout_svg_params::color_background,
                        DOC(fiction_write_sidb_layout_svg_params_color_background));
 
-    detail::write_sidb_layout_svg_impl<py_charge_distribution_surface_100>(m);
-    detail::write_sidb_layout_svg_impl<py_charge_distribution_surface_100>(m);
-    detail::write_sidb_layout_svg_impl<py_charge_distribution_surface>(m);
+    py::class_<fiction::write_qca_layout_svg_params>(m, "write_qca_layout_svg_params",
+                                                     DOC(fiction_write_qca_layout_svg_params))
+        .def(py::init<>())
+        .def_readwrite("simple", &fiction::write_qca_layout_svg_params::simple,
+                       DOC(fiction_write_qca_layout_svg_params_simple));
+    ;
 
-    detail::write_sidb_layout_svg_impl<py_sidb_111_lattice>(m);
-    detail::write_sidb_layout_svg_impl<py_sidb_100_lattice>(m);
-    detail::write_sidb_layout_svg_impl<py_sidb_layout>(m);
+    detail::write_svg_layout_impl<py_charge_distribution_surface_100>(m);
+    detail::write_svg_layout_impl<py_charge_distribution_surface_100>(m);
+    detail::write_svg_layout_impl<py_charge_distribution_surface>(m);
+
+    detail::write_svg_layout_impl<py_sidb_111_lattice>(m);
+    detail::write_svg_layout_impl<py_sidb_100_lattice>(m);
+    detail::write_svg_layout_impl<py_sidb_layout>(m);
 }
 
 }  // namespace pyfiction
 
-#endif  // PYFICTION_WRITE_SIDB_LAYOUT_SVG_HPP
+#endif  // PYFICTION_WRITE_SVG_LAYOUT_HPP
