+++ bindings/pyfiction/pyfiction.cpp
@@ -61,15 +61,15 @@
 #include "pyfiction/inout/write_dot_layout.hpp"
 #include "pyfiction/inout/write_fgl_layout.hpp"
 #include "pyfiction/inout/write_fqca_layout.hpp"
-#include "pyfiction/inout/write_sidb_layout_svg.hpp"
+#include "pyfiction/inout/write_svg_layout.hpp"
 // #include "pyfiction/inout/write_location_and_ground_state.hpp"
 #include "pyfiction/inout/write_operational_domain.hpp"
 #include "pyfiction/inout/write_qca_layout.hpp"
-#include "pyfiction/inout/write_qca_layout_svg.hpp"
 #include "pyfiction/inout/write_qcc_layout.hpp"
 #include "pyfiction/inout/write_qll_layout.hpp"
 #include "pyfiction/inout/write_sqd_layout.hpp"
 #include "pyfiction/inout/write_sqd_sim_result.hpp"
+#include "pyfiction/inout/write_svg_layout.hpp"
 #include "pyfiction/layouts/cartesian_layout.hpp"
 #include "pyfiction/layouts/cell_level_layout.hpp"
 #include "pyfiction/layouts/clocked_layout.hpp"
@@ -217,7 +217,6 @@ PYBIND11_MODULE(pyfiction, m, pybind11::mod_gil_not_used())
     pyfiction::write_sqd_sim_result(m);
     // pyfiction::write_location_and_ground_state(m);
     pyfiction::write_operational_domain(m);
-    pyfiction::write_sidb_layout_svg(m);
     pyfiction::read_fgl_layout(m);
     pyfiction::read_fqca_layout(m);
     pyfiction::read_sqd_layout_100(m);
