+++ docs/io/visualization.rst
@@ -50,8 +50,6 @@ SVG Images
 
         .. doxygenclass:: fiction::unsupported_cell_type_exception
 
-        **Header:** ``fiction/io/write_sidb_layout_svg.hpp``
-
         .. doxygenstruct:: fiction::color_mode
            :members:
 
@@ -63,6 +61,14 @@ SVG Images
         .. doxygenfunction:: fiction::write_sidb_layout_svg(const Lyt& lyt, const std::string_view& filename,
                                              const write_sidb_layout_svg_params& ps = {})
 
+
+        .. doxygenstruct:: fiction::write_svg_layout_params
+           :members:
+
+        .. doxygenfunction:: fiction::write_svg_layout(const Lyt& lyt, std::ostream& os, const write_svg_layout_params& ps = {})
+        .. doxygenfunction:: fiction::write_svg_layout(const Lyt& lyt, const std::string_view& filename,
+                           const write_svg_layout_params& ps = {})
+
     .. tab:: Python
         .. autoclass:: mnt.pyfiction.write_qca_layout_svg_params
             :members:
@@ -71,4 +77,4 @@ SVG Images
         .. autoclass:: mnt.pyfiction.write_sidb_layout_svg_params
             :members:
         .. autofunction:: mnt.pyfiction.write_sidb_layout_svg
-        .. autofunction:: mnt.pyfiction.write_sidb_layout_svg_and_return_as_string
+        .. autofunction:: mnt.pyfiction.write_sidb_layout_svg_to_string
